+++ website/src/routes/publish.lazy.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import { ConnectKitButton, useModal } from 'connectkit';
+import { graphql } from 'gql.tada';
 import { useForm } from 'react-hook-form';
 import semver from 'semver';
 import { Address } from 'viem';
@@ -27,20 +28,30 @@ import {
 } from '@/components/ui/select';
 import { Textarea } from '@/components/ui/textarea';
 import { useToast } from '@/components/ui/use-toast';
+import { CHAINS } from '@/lib/constants';
+import {
+  NETWORK_SUBGRAPH_MAINNET,
+  NETWORK_SUBGRAPH_SEPOLIA,
+  networkSubgraphExecute,
+} from '@/lib/graphql';
 import { readIpfsFile, uploadFileToIpfs } from '@/lib/ipfs';
 import { ipfsHexHash } from '@/lib/utils';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useQuery } from '@tanstack/react-query';
-import { createFileRoute } from '@tanstack/react-router';
+import { createLazyFileRoute } from '@tanstack/react-router';
 import { L2GNSABI } from '../abis/L2GNS';
 import addresses from '../addresses.json';
 
-const CHAINS = ['arbitrum-one', 'arbitrum-sepolia'] as const;
 const SUPPORTED_CHAIN = {
-  'arbitrum-one': { chainId: 42161, contracts: addresses[42161] },
+  'arbitrum-one': {
+    chainId: 42161,
+    contracts: addresses[42161],
+    subgraph: NETWORK_SUBGRAPH_MAINNET,
+  },
   'arbitrum-sepolia': {
     chainId: 421614,
     contracts: addresses[421614],
+    subgraph: NETWORK_SUBGRAPH_SEPOLIA,
   },
 } as const;
 
@@ -118,15 +129,61 @@ const Manifest = z.object({
   repository: z.string().describe('An optional link to where the subgraph lives.').optional(),
 });
 
-function DeploySubgraph({ deploymentId }: { deploymentId: string }) {
-  const { writeContractAsync, isPending } = useWriteContract({});
+const GetSubgraphInfo = graphql(/* GraphQL */ `
+  query GetSubgraphInfo($subgraphId: ID!) {
+    subgraph(id: $subgraphId) {
+      id
+      owner {
+        id
+      }
+      nftID
+      metadata {
+        displayName
+        image
+        description
+        website
+        codeRepository
+      }
+      versions {
+        id
+        subgraphDeployment {
+          ipfsHash
+        }
+        metadata {
+          label
+        }
+      }
+    }
+  }
+`);
+
+function getEtherscanUrl({ chainId, hash }: { chainId: number; hash: string }) {
+  switch (chainId) {
+    case SUPPORTED_CHAIN['arbitrum-one'].chainId:
+      return `https://arbiscan.io/tx/${hash}`;
+    case SUPPORTED_CHAIN['arbitrum-sepolia'].chainId:
+      return `https://sepolia.arbiscan.io/tx/${hash}`;
+    default:
+      throw new Error('Invalid chain');
+  }
+}
+
+function DeploySubgraph({
+  deploymentId,
+  subgraphId,
+  network,
+}: {
+  deploymentId: string;
+  subgraphId: string | undefined;
+  network: (typeof CHAINS)[number] | undefined;
+}) {
+  const { writeContractAsync, isPending, error: contractError } = useWriteContract({});
   const { setOpen } = useModal();
   const { switchChainAsync, isPending: chainSwitchPending } = useSwitchChain();
   const { toast } = useToast();
   const [deployed, setDeployed] = useState(false);
 
   const { address, chainId } = useAccount();
-
   const { data: subgraphManifest } = useQuery({
     queryKey: ['subgraph-manifest', deploymentId],
     queryFn: async () => {
@@ -150,8 +207,91 @@ function DeploySubgraph({ deploymentId }: { deploymentId: string }) {
       codeRepository: subgraphManifest?.parsed.repository,
       website: undefined,
       categories: undefined,
+      chain: network,
     },
   });
+
+  const chain = form.watch('chain');
+
+  const { data: subgraphInfo } = useQuery({
+    queryKey: ['subgraph-info', subgraphId, chain, chainId],
+    queryFn: async () => {
+      if (!subgraphId) return;
+
+      const subgraphEndpoint = chain ? getChainInfo(chain)?.subgraph : null;
+
+      if (!subgraphEndpoint) return;
+
+      const data = await networkSubgraphExecute(
+        GetSubgraphInfo,
+        { subgraphId },
+        {
+          endpoint: subgraphEndpoint,
+        },
+      );
+
+      const metadata = data.subgraph?.metadata;
+      if (metadata) {
+        if (metadata.image) {
+          form.setValue('subgraphImage', metadata.image);
+        }
+
+        if (metadata.description) {
+          form.setValue('description', metadata.description);
+        }
+
+        if (metadata.codeRepository) {
+          form.setValue('codeRepository', metadata.codeRepository);
+        }
+
+        if (metadata.website) {
+          form.setValue('website', metadata.website);
+        }
+
+        if (metadata.displayName) {
+          form.setValue('displayName', metadata.displayName);
+        }
+      }
+
+      return data;
+    },
+    enabled: !!subgraphId && !!chain,
+  });
+
+  function updateMetadata() {
+    if (!subgraphInfo) return false;
+
+    const version = form.watch('versionLabel');
+
+    const versionInfo = subgraphInfo.subgraph?.versions.find(
+      ({ metadata }) => metadata?.label === version,
+    );
+
+    if (!versionInfo) return false;
+
+    return versionInfo.subgraphDeployment.ipfsHash === deploymentId;
+  }
+
+  function ensureNewVersion() {
+    // If there is no subgraph ID then it means we are just deploying a new subgraph
+    if (!subgraphId) return true;
+
+    if (!subgraphInfo) return false;
+
+    const version = form.watch('versionLabel');
+
+    return !subgraphInfo.subgraph?.versions.some(({ metadata }) => metadata?.label === version);
+  }
+
+  function isOwner() {
+    // If there is no subgraph ID then it means we are just deploying a new subgraph
+    if (!subgraphId) return true;
+
+    // we expect to have the subgraph info
+    if (!subgraphInfo) return false;
+
+    return address?.toString().toLowerCase() === subgraphInfo.subgraph?.owner.id?.toLowerCase();
+  }
 
   async function onSubmit(values: z.infer<typeof subgraphMetadataSchema>) {
     const selectedChain = getChainInfo(values.chain);
@@ -174,6 +314,83 @@ function DeploySubgraph({ deploymentId }: { deploymentId: string }) {
       content: Buffer.from(JSON.stringify({ label: values.versionLabel, description: null })),
     });
 
+    if (subgraphId) {
+      if (!subgraphInfo?.subgraph?.nftID) {
+        toast({
+          description:
+            "Missing Subgraph ID from subgraph. Try again, if it still doesn't work, please contact support.",
+          variant: 'destructive',
+        });
+        return;
+      }
+
+      // we are just updating the metadata
+      if (updateMetadata() && !ensureNewVersion()) {
+        const subgraphMeta = await uploadFileToIpfs({
+          path: '',
+          content: Buffer.from(
+            JSON.stringify(
+              subgraphMetadata({
+                ...values,
+              }),
+            ),
+          ),
+        });
+
+        const hash = await writeContractAsync({
+          address: selectedChain.contracts.L2GNS.address as Address,
+          abi: L2GNSABI,
+          functionName: 'updateSubgraphMetadata',
+          args: [BigInt(subgraphInfo.subgraph.nftID), ipfsHexHash(subgraphMeta)],
+        });
+
+        const etherscanUrl = getEtherscanUrl({ chainId: selectedChain.chainId, hash });
+
+        window.open(etherscanUrl, '_blank');
+        setDeployed(true);
+        toast({
+          description: 'You are all set! You can go back to the CLI and close this window',
+        });
+
+        return;
+      }
+
+      try {
+        const hash = await writeContractAsync({
+          address: selectedChain.contracts.L2GNS.address as Address,
+          abi: L2GNSABI,
+          functionName: 'publishNewVersion',
+          args: [BigInt(subgraphInfo.subgraph.nftID), DEPLOYMENT_ID, ipfsHexHash(versionMeta)],
+        });
+
+        const etherscanUrl = getEtherscanUrl({ chainId: selectedChain.chainId, hash });
+
+        window.open(etherscanUrl, '_blank');
+        setDeployed(true);
+        toast({
+          description: 'You are all set! You can go back to the CLI and close this window',
+        });
+      } catch (err) {
+        const e = contractError;
+        if (e?.name === 'ContractFunctionExecutionError') {
+          if (e.cause.name === 'ContractFunctionRevertedError') {
+            toast({
+              description: e.cause.reason,
+              variant: 'destructive',
+            });
+            return;
+          }
+        }
+
+        toast({
+          description:
+            'An error occurred while trying to deploy the new version. Please try again later or contact support',
+          variant: 'destructive',
+        });
+      }
+      return;
+    }
+
     const subgraphMeta = await uploadFileToIpfs({
       path: '',
       content: Buffer.from(
@@ -192,20 +409,30 @@ function DeploySubgraph({ deploymentId }: { deploymentId: string }) {
       args: [DEPLOYMENT_ID, ipfsHexHash(versionMeta), ipfsHexHash(subgraphMeta)],
     });
 
-    window.open(`https://sepolia.arbiscan.io/tx/${hash}`, '_blank');
+    const etherscanUrl = getEtherscanUrl({ chainId: selectedChain.chainId, hash });
+
+    window.open(etherscanUrl, '_blank');
     setDeployed(true);
     toast({
       description: 'You are all set! You can go back to the CLI and close this window',
     });
+    return;
   }
 
   const isDeployButtonDisabled =
-    deployed || chainSwitchPending || isPending || !form.formState.isValid;
+    deployed ||
+    chainSwitchPending ||
+    isPending ||
+    !form.formState.isValid ||
+    !(ensureNewVersion() || updateMetadata()) ||
+    !isOwner();
 
   const deployButtonCopy = (() => {
+    if (!address) return 'Need to connect wallet';
     if (deployed) return 'Deployed';
     if (chainSwitchPending) return 'Switching Chains...';
     if (isPending) return 'Check Wallet...';
+    if (updateMetadata()) return 'Update Metadata';
     return 'Deploy';
   })();
 
@@ -338,7 +565,14 @@ function DeploySubgraph({ deploymentId }: { deploymentId: string }) {
 }
 
 function Page() {
-  const { id } = Route.useSearch();
+  const { id, subgraphId, network } = Route.useSearch();
+
+  const protocolNetwork = network
+    ? // @ts-expect-error we want to compare if it is a string or not
+      CHAINS.includes(network.toLowerCase())
+      ? (network.toLowerCase() as (typeof CHAINS)[number])
+      : undefined
+    : undefined;
 
   return (
     <div className="min-h-screen">
@@ -347,7 +581,7 @@ function Page() {
         <ConnectKitButton />
       </nav>
       {id ? (
-        <DeploySubgraph deploymentId={id} />
+        <DeploySubgraph deploymentId={id} subgraphId={subgraphId} network={protocolNetwork} />
       ) : (
         <div className="flex justify-center items-center min-h-screen -mt-16">
           Unable to find the Deployment ID. Go back to CLI
@@ -357,7 +591,6 @@ function Page() {
   );
 }
 
-export const Route = createFileRoute('/publish')({
+export const Route = createLazyFileRoute('/publish')({
   component: Page,
-  validateSearch: z.object({ id: z.string() }),
 });